renaming optimism styled components
